+++ src/main/java/com/example/springboottemplate/interceptor/AuthInterceptor.java
@@ -53,7 +53,7 @@ public class AuthInterceptor implements HandlerInterceptor {
 
         // 5. 解析token并将用户信息存入request
         Claims claims = jwtUtil.parseToken(token);
-        Integer userId = jwtUtil.getUserId(claims);
+        Long userId = jwtUtil.getUserId(claims);
         String username = claims.getSubject();
 
         request.setAttribute("userId", userId);
