+++ src/main/java/com/example/springboottemplate/mapper/system/UserMapper.java
@@ -25,5 +25,5 @@ public interface UserMapper extends BaseMapper<User> {
     void updateUser(User user); //修改用户信息
 
     /** 解除微信身份并逻辑删除（合并账号时清理临时 wx 用户） */
-    void detachWxAndSoftDelete(Integer id);
+    void detachWxAndSoftDelete(Long id);
 }
